# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Contacts API",
      "version": "0.0.1",
      "description": "REST CONTACTS API (WDD341) by Greg Azores",
      "termsOfService": "http://swagger.io/terms/"
    },
    "servers": [ 
        {
        "url": "http://localhost:3000"
        }
    ],
    "basePath": "/",
    "schemes": [
      "http",
      "https"
    ],
...
    "paths": {
...
      "/users/register/": {
        "post": {
          "tags": ["user-login"],
          "description": "Here is where user will register username and password.",
          "parameters": [
              {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "schema": {
                      "$ref": "#/components/schemas/user"
                  }
                      
              }
          ],
          "responses": {}
        }
      },
      "/users/login/": {
        "post": {
          "tags": ["user-login"],
          "description": "Here is where user will login username and password.",
          "parameters": [
              {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "schema": {
                      "$ref": "#/components/schemas/user"
                  }
                      
              }
          ],
          "responses": {}
        }
      },
...
    },
    "components": {
      "schemas": {
...
          "user": {
            "type": "object",
            "properties": {
                "username": {
                    "example": "greg azores"
                },

                "password": {
                    "example": "password"
                }
            }
        }
      }
  }
  }